requirements.txt

fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
google-genai>=0.8.0
supabase>=2.0.0
pydantic>=2.0.0
speechrecognition>=3.10.0
pyautogui>=0.9.54
pywin32>=306; platform_system == "Windows"